* kandinsky compatibility fix * revo error layout fix * lang fix

+++ inc/leyka-core.php
@@ -896,8 +896,14 @@ class Leyka extends Leyka_Singleton {
         }
 
         if( !$leyka_last_ver || $leyka_last_ver < '3.0' ) {
-
-            update_option('leyka_init_wizard_redirect', !$leyka_last_ver);
+            
+            if(class_exists( 'TGM_Plugin_Activation' )) {
+              update_option('leyka_init_wizard_redirect', false);
+            }
+            else {
+               update_option('leyka_init_wizard_redirect', !$leyka_last_ver);
+            }
+            
             update_option('leyka_receiver_country', 'ru');
             update_option('leyka_receiver_legal_type', 'legal');
